Refactor PropagationLayer to use float32 for inputs; remove outdated files and add new outputs
- Changed input casting in PropagationLayer from float16 to float32 for improved precision.
- Deleted several outdated BMP and PNG files from best_doe_masks and sample_outputs directories.
- Added new output files for PSNR 24.19 and 31.10 in sample_outputs and saved_histories.
- Removed old history graphs and STL model files that are no longer needed.

+++ cd2nn_power/PropagationLayer.py
@@ -68,7 +68,7 @@ class PropagationLayer(tf.keras.layers.Layer):
         )
 
     def call(self, inputs):
-        inputs = tf.cast(inputs, tf.float16)
+        inputs = tf.cast(inputs, tf.float32)
         re_u = inputs[..., 0]
         im_u = inputs[..., 1]
